requirements.txt
flask
mysql-connector-python
flask-mysqldb
PyJWT
flask_mysqldb
mysql.connector
flask-cors